% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{GetData}
\alias{GetData}
\title{Get essential data}
\usage{
GetData(exampleData)
}
\arguments{
\item{exampleData}{A character, should be one of "DrugSPCentralMatrix", "DrugSPESMatrix", "DrugSPESCMatrix", "DrugSPPvalueMatrix", "GO_SP", "Subpathways", "GSE65194expr" and "label".}
}
\value{
Data
}
\description{
This function is used to achieve essential data.
}
